estimation-game/backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
psycopg2-binary==2.9.9
redis==5.0.3
python-decouple==3.8
